+++ src/storybook/utils/index.ts
@@ -1,3 +1,2 @@
 export * from './loaders';
-export * from './utils';
 export * from './StoryContainer';
